Trigger deferred test on retrieving result
In our case we chill our selenium UI tests. As a feature a screenshot is created when the test fails. Currently this is only possible by including the given and when inside the actual test method (then) because when an exception occurs in the constructor it is not possible to catch it for building the screenshot.

Solution would be use a deffered test, but that is currently only possible when expecting an exception. This is now supported through the Result property which will execute the deferred test first.